Model2/requirements.txt
# Core ML and Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
keras==2.13.1
torch==2.0.1
torchvision==0.15.2

# Time Series Analysis
statsmodels==0.14.0
prophet==1.1.4
pmdarima==2.0.3

# Geospatial Processing
geopandas==0.13.2
shapely==2.0.1
folium==0.14.0
geopy==2.3.0

# API and Web Framework
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
requests==2.31.0

# Database
pymongo==4.4.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration and Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
schedule==1.2.0
logging==0.4.9.6

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Data Sources
tweepy==4.14.0
beautifulsoup4==4.12.2

# Model Serving
joblib==1.3.2
pickle-mixin==1.0.2